chore(project): separate containers from components
