+++ ui/src/components/TechBadge.tsx
