+++ ui/src/components/customIcons/LogoIcon.tsx
