+++ ui/src/containers/ProjectFilters.tsx
@@ -2,7 +2,7 @@
 import { useState, useEffect } from 'react';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { Search, X } from 'lucide-react';
-import { TechBadge } from './TechBadge';
+import { TechBadge } from '../components/TechBadge';
 
 export interface ProjectFiltersProps {
   allTech: string[];
